Setup for multi-source project
rmkit is header-only, but it includes the full implementation (without
guards) . . . so everything that uses rmkit needs to be in a single
translation unit. Until there's a better solution for that, this just
cats all the cpp files together. So far builds aren't so bad with this
method, but we'll see as the project grows.