Add custom colors for a handful of games
These are mostly based on the excellent adjustments made in
PocketPuzzles. PocketPuzzles edited the game source to change the number
of colors in some games (e.g. to remove flashing), but for now I'm going
to leave the puzzle source alone, so I had to make some changes to
accommodate the extra colors in the original.