Fix bug with canvas drawing outside its bounds
When the game drawing area is smaller than the size of the canvas, the
virtual fb is translated so that the game is centered. In this case, the
"full width" of the game area is the fb size minus the translation area.

Note that in the case of a perfectly centered game, the dirty rect does
still extend to the edge of the _canvas_ (this larger than the actual
game area), but the bug was caused by the canvas trying to render
outside the bounds of the canvas widget itself.