Flash to clear ghosting on scene changes
I'm seeing a lot of ghosting when going from black -> gray, and this
seems to take care of it. This is used in two places:

1. After closing the help dialog, since text is particularly conspicuous
   when switching to a gray background.
2. After loading a new game, since there's often a lot of state to clear
   from the old game (or the chooser screen if that's where we're coming
   from).

This does not take care of ghosting between moves (e.g. black box, when
removing one of the balls), nor does it handle the generic case of
hiding an overlay, like the presets menu. Should probably hook into
ui::MainLoop for that.

This is only useful for games that draw many shades of gray. For games
that are purely black and white, this is unnecessary, and is probably
distracting. It might be worth making this configurable.