When flashing canvas, assume game is centered
Rather than exploring the entire convas bounding box for the edge of the
actual drawing, start with just the center (as determined from
trans_x/trans_y).

Without this, the flash is extended to cover the controls text when
there's a toast shown, since that text shows up as black pixels.